Add loaders for proposal expiry and session updated

+++ protocol/sign/src/main/kotlin/com/walletconnect/sign/engine/use_case/calls/RejectSessionUseCase.kt
@@ -3,12 +3,16 @@ package com.walletconnect.sign.engine.use_case.calls
 import com.walletconnect.android.internal.common.model.IrnParams
 import com.walletconnect.android.internal.common.model.Tags
 import com.walletconnect.android.internal.common.model.type.JsonRpcInteractorInterface
+import com.walletconnect.android.internal.common.scope
 import com.walletconnect.android.internal.common.storage.verify.VerifyContextStorageRepository
+import com.walletconnect.android.internal.utils.CoreValidator.isExpired
 import com.walletconnect.android.internal.utils.FIVE_MINUTES_IN_SECONDS
 import com.walletconnect.foundation.common.model.Ttl
 import com.walletconnect.sign.common.exceptions.PeerError
+import com.walletconnect.sign.common.exceptions.SessionProposalExpiredException
 import com.walletconnect.sign.engine.model.mapper.toSessionProposeRequest
 import com.walletconnect.sign.storage.proposal.ProposalStorageRepository
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.supervisorScope
 
 internal class RejectSessionUseCase(
@@ -19,14 +23,21 @@ internal class RejectSessionUseCase(
 
     override suspend fun reject(proposerPublicKey: String, reason: String, onSuccess: () -> Unit, onFailure: (Throwable) -> Unit) = supervisorScope {
         val proposal = proposalStorageRepository.getProposalByKey(proposerPublicKey)
-        proposalStorageRepository.deleteProposal(proposerPublicKey)
-        verifyContextStorageRepository.delete(proposal.requestId)
+        if (proposal.expiry.isExpired()) {
+            throw SessionProposalExpiredException("Session proposal expired")
+        }
 
         jsonRpcInteractor.respondWithError(
             proposal.toSessionProposeRequest(),
             PeerError.EIP1193.UserRejectedRequest(reason),
             IrnParams(Tags.SESSION_PROPOSE_RESPONSE, Ttl(FIVE_MINUTES_IN_SECONDS)),
-            onSuccess = { onSuccess() },
+            onSuccess = {
+                scope.launch {
+                    proposalStorageRepository.deleteProposal(proposerPublicKey)
+                    verifyContextStorageRepository.delete(proposal.requestId)
+                }
+                onSuccess()
+            },
             onFailure = { error -> onFailure(error) })
     }
 }
